Implement Meson build system
This is intended as a replacement to the current CMake build system.
Mostly because the alterations I had to make, however small they may be,
currently break builds in CMake without manual intervention.

Also, VCPKG support is missing, mostly due to my inexperience on this
side of the equation, and the origins of this PR being my own
frustrations with VCPKG & CMake.

I attempted to make as few changes as possible, but some still had to be
made:

- `src/unittest.cpp` now reads `Robot.pcm` from a different
path. (Funnily enough, it's the path that's actually documented in
`docpages/example_programs/music_and_audio/soundboard.md`.)

- `src/dpp/utility.cpp`, `src/dpp/discordclient.cpp`, and
`src/dpp/discordvoiceclient.cpp` now include `config.h`, a file
generated by Meson. (This is why CMake builds are broken.)

*one cherry pick later...*

Now filed & retested against the dev branch.