Use string views
This will make the library easier to use for std::string_view users as a
LOT of explicit std::string creations are needed atm. There should be
some performance improvements as well (less allocations, blablabla).

This is mostly backwards compatible, but will cause issues with implicit
conversions. For example, within the library itself, there were some
assignments of a utility::icon to a string that would implicitly create
a utility::iconhash - now it does not, and does not compile without
explicitly creating the iconhash. This could be accounted for with a bit
more code, or just reverting the usage of std::string_view, if it's
important enough.

There's also some new stuff / changes outside of just changing
std::strings to std::string_views:
- A new utility function for maps, dpp::utility::emplace_or_assign,
  which essentially is insert_or_assign but with the in-place
  construction that emplace offers.
- All non-floating point instances of std::stoX that would take in a
  std::string_view have been changed to use std::from_chars.
- 2 new types, dpp::membuf and dpp::imemstream, in stringops.h, which
  allow input streams over a raw character array, necessary for
  std::string_view. These are not compiled if std::ispanstream is
  available, which was added in C++23, as it is far superior. All
  instances of std::istringstream that would take in a std::string_view
  have been replaced to use one or the other, depending on what's
  available.